Renamed mcu's to platformio board names
