+++ mirte_telemetrix/config/old/mirte_pcb06_config.yaml
